+++ bin/settings.pbjson.dart
@@ -1,0 +1,33 @@
+// This is a generated file - do not edit.
+//
+// Generated from settings.proto.
+
+// @dart = 3.3
+
+// ignore_for_file: annotate_overrides, camel_case_types, comment_references
+// ignore_for_file: constant_identifier_names
+// ignore_for_file: curly_braces_in_flow_control_structures
+// ignore_for_file: deprecated_member_use_from_same_package, library_prefixes
+// ignore_for_file: non_constant_identifier_names, prefer_relative_imports
+// ignore_for_file: unused_import
+
+import 'dart:convert' as $convert;
+import 'dart:core' as $core;
+import 'dart:typed_data' as $typed_data;
+
+@$core.Deprecated('Use settingsDescriptor instead')
+const Settings$json = {
+  '1': 'Settings',
+  '2': [
+    {'1': 'name', '3': 1, '4': 1, '5': 9, '8': {}, '10': 'name'},
+    {'1': 'retries', '3': 2, '4': 1, '5': 5, '8': {}, '10': 'retries'},
+    {'1': 'region', '3': 3, '4': 1, '5': 9, '8': {}, '10': 'region'},
+    {'1': 'verbose', '3': 4, '4': 1, '5': 8, '8': {}, '10': 'verbose'},
+  ],
+};
+
+/// Descriptor for `Settings`. Decode as a `google.protobuf.DescriptorProto`.
+final $typed_data.Uint8List settingsDescriptor = $convert.base64Decode(
+    'CghTZXR0aW5ncxIXCgRuYW1lGAEgASgJQgOQUgFSBG5hbWUSHgoHcmV0cmllcxgCIAEoBUIEml'
+    'IBM1IHcmV0cmllcxIkCgZyZWdpb24YAyABKAlCDJpSCXVzLWVhc3QtMVIGcmVnaW9uEiEKB3Zl'
+    'cmJvc2UYBCABKAhCB5pSBHRydWVSB3ZlcmJvc2U=');